requirements.txt
pytest
pytest-html
allure-pytest
PyYAML
selenium
pytest-rerunfailures
